v0.0.1 initial commit : basic functionalities
This inital version contains a hook receiver which is able to parse
a gitlab payload and detect any references to issues.
It will then add a reference to the commit in the journal history
of the related issue.
This commit also contains some basic configuration like choosing
which user is linked to the journal entry (custom bots).